Fix organization logos
This gets the organization logos, linked to the organization's site and
with the organization's title as alt text, working on organization and
dataset read pages.

Still need to fix other organization and dataset pages like edit,
history, etc.

Need to add more careful tests for this.

Need to run CKAN tests to check that this doesn't break any.