Make column titles on dataset listing clickable
If you click on the 'Name' or 'Popularity' column titles at the top of
the list of datasets on the front page, this will reload the page with
the datasets sorted by name or popularity. Clicking again will reverse
the sort order.

Currently working for name but not popularity.

Also the old 'Order by' dropdown menu is still there, should be removed
if we get the new links all working.